requirements.txt
pytest==7.1.2
allure-pytest==2.10.0
requests==2.28.1
jsonpath==0.82
psycopg2-binary==2.9.3


